refactor: rename classes for improved clarity and consistency across tests
