+++ test/qubic.test.ts
@@ -1,4 +1,4 @@
-import Qubic from "../src/qubic";
+import Qubic from "../src/core";
 
 describe("Qubic SDK Structure", () => {
   let qubic: Qubic;
