+++ test/utils.test.ts
@@ -1,8 +1,8 @@
-import { Utils } from "../src/utils";
+import { UtilityService } from "../src/utils";
 
-describe("Utils Module", () => {
+describe("UtilityService Module", () => {
   test("should create a random seed", () => {
-    const seed = Utils.createSeed();
+    const seed = UtilityService.createSeed();
     expect(seed).toBeDefined();
     expect(seed).toHaveLength(55);
     expect(seed).toMatch(/^[a-z]{55}$/);
@@ -10,7 +10,7 @@ describe("Utils Module", () => {
 
   test("should convert bytes to shifted hex", () => {
     const bytes = new Uint8Array([15, 240, 45, 200]);
-    const hex = Utils.bytesToShiftedHex(bytes);
+    const hex = UtilityService.bytesToShiftedHex(bytes);
     expect(hex).toBeDefined();
     expect(typeof hex).toBe("string");
     expect(hex.length).toBe(bytes.length * 2);
@@ -19,49 +19,49 @@ describe("Utils Module", () => {
   test("should convert valid public key string to bytes", () => {
     const publicKeyString =
       "ABCDEFGHIJKLMNOPQRSTUVWXYZABCDEFGHIJKLMNOPQRSTUVWXYZABCD";
-    const bytes = Utils.publicKeyStringToBytes(publicKeyString);
+    const bytes = UtilityService.publicKeyStringToBytes(publicKeyString);
     expect(bytes).toBeInstanceOf(Uint8Array);
     expect(bytes.length).toBe(32);
   });
 
   test("should convert bytes32 to string", () => {
     const bytes = new Uint8Array(32);
-    const str = Utils.bytes32ToString(bytes);
+    const str = UtilityService.bytes32ToString(bytes);
     expect(typeof str).toBe("string");
     expect(str.length).toBe(56);
   });
 
   test("should convert digest bytes to string", () => {
     const bytes = new Uint8Array(32);
-    const digestStr = Utils.digestBytesToString(bytes);
+    const digestStr = UtilityService.digestBytesToString(bytes);
     expect(typeof digestStr).toBe("string");
     expect(digestStr.length).toBe(56);
   });
 
   test("should convert public key bytes to string", () => {
     const bytes = new Uint8Array(32);
-    const publicKeyStr = Utils.publicKeyBytesToString(bytes);
+    const publicKeyStr = UtilityService.publicKeyBytesToString(bytes);
     expect(typeof publicKeyStr).toBe("string");
     expect(publicKeyStr.length).toBe(56);
   });
 
   test("should convert valid seed string to bytes", () => {
     const seed = "abcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxy";
-    const bytes = Utils.seedStringToBytes(seed);
+    const bytes = UtilityService.seedStringToBytes(seed);
     expect(bytes).toBeInstanceOf(Uint8Array);
     expect(bytes.length).toBe(seed.length);
   });
 
   test("should convert hex string to base26", () => {
     const hex = "0xC4aE1B5f8872D961cF6C1c9E16A3FCAe2C89e5E6";
-    const base26 = Utils.hexToBase26(hex);
+    const base26 = UtilityService.hexToBase26(hex);
     expect(base26).toBeDefined();
     expect(typeof base26).toBe("string");
   });
 
   test("should convert base26 to hex string", () => {
     const base26 = "wslioomgijdjtuhrjvoswvwgjczbnikuve";
-    const hex = Utils.base26ToHex(base26);
+    const hex = UtilityService.base26ToHex(base26);
     expect(hex).toBeDefined();
     expect(typeof hex).toBe("string");
   });
